+++ src/api/SpacefillAPIWrapperV1.ts
@@ -65,7 +65,7 @@ export class SpacefillAPIWrapperV1 {
           ...APIContext.getWorkflowHeader(this.workflowType),
         },
         transformRequest: (data, headers) => {
-          headers["Spacefill-Ctx-Data-Source"] = this.dataSource ?? "unknown";
+          headers["Spacefill-Ctx-Data-Source"] = this.dataSource?.substring(0, 100) ?? "unknown";
           if (!headers["Content-Type"]) {
             headers["Content-Type"] = "application/json";
           }