+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.6] - 2023-12-06
+
+### Fixed
+
+- [#15](https://gitlab.spacefill.fr/spacefill/integration-framework/-/issues/15) Limit length of `Spacefill-Ctx-Data-Source` header to 100 characters.
+
 ## [0.0.5] - 2023-12-05
 
 ### Added